backend/requirements.txt
fastapi==0.115.6
loguru==0.7.3
matplotlib==3.8.4
orjson==3.10.14
packaging==24.0
pandas==2.2.2
plotly==6.0.1
requests==2.32.3
uvicorn==0.34.0
sqlalchemy==2.0.40
asyncpg==0.30.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1